refactor: move client config details to application variable
